[redux][m]: introduce to Redux store. Mocked store moved into index.js and passed as propos. CountryPerformanceOnRisk state set via mapStateToProps
